Implement auto-update function
The bot will automatically attempt to update itself if enabled via the management portal. This is a very new system and may be prone to bugs initially. Full-project backups are created during updates.

The bot checks for configuration updates before loading deferred configs on initialization.

Remove update.json file, as it is no longer used. All update checks utilize the GitHub API to get information on the latest release.