Cache custom command data on start
Custom command data is cached on start to improve performance when using custom commands. My testing showed a 7.06% improvement over five runs in this commit.

There is still a lot to be done regarding Osmium. It will require a significant overhaul to improve its management of resources. Launching a multiprocessing manager on each instance takes a significant amount of time.