Osmium rewrite
Rewrite Osmium with performance in mind. Modify the custom commands caching behavior with a custom cache loader, as the typical reload method won't work well here.

Testing shows that compared to the legacy approach, performance has increased 127.66% with the rewrite and the caching features. Execution time for a simple command is on average, roughly a second.