Removal of unused features
Remove custom commands system as it is mostly unused and a heavy maintenance burden.

Remove database cache manager system. It was only used by the custom commands system to preload commands and no longer has a purpose as most content is stored on the management portal.

Remove data migration system. Irrelevant with a major rewrite, and the only thing still saved guild-specific locally is configurations, which should just be kept in memory and fetched as needed.